move safe container from components
