+++ src/test/safe.dom.funds.thresholdGt1.ts
@@ -9,7 +9,7 @@ import { sleep } from 'src/utils/timer'
 import '@testing-library/jest-dom/extend-expect'
 import { BALANCE_ROW_TEST_ID } from 'src/routes/safe/components/Balances'
 import { fillAndSubmitSendFundsForm } from './utils/transactions'
-import { TRANSACTIONS_TAB_BTN_TEST_ID } from 'src/routes/safe/components/Layout/index'
+import { TRANSACTIONS_TAB_BTN_TEST_ID } from 'src/routes/safe/container'
 import { TRANSACTION_ROW_TEST_ID } from 'src/routes/safe/components/Transactions/TxsTable'
 import { useTestAccountAt, resetTestAccount } from './utils/accounts'
 //import { CONFIRM_TX_BTN_TEST_ID } from 'src/routes/safe/components/Transactions/TxsTable/ExpandedTx/OwnersColumn/ButtonRow'
