+++ src/test/utils/transactions/transactionList.helper.ts
@@ -2,7 +2,7 @@
 import { fireEvent } from '@testing-library/react'
 import { sleep } from 'src/utils/timer'
 import { shortVersionOf } from 'src/logic/wallets/ethAddresses'
-import { TRANSACTIONS_TAB_BTN_TEST_ID } from 'src/routes/safe/components/Layout/index'
+import { TRANSACTIONS_TAB_BTN_TEST_ID } from 'src/routes/safe/container'
 import { TRANSACTION_ROW_TEST_ID } from 'src/routes/safe/components/Transactions/TxsTable'
 import {
   TRANSACTIONS_DESC_ADD_OWNER_TEST_ID,
